Made a new dual extruder mount
extruders+induction sensor :+1:
This is designed to mount on either of the two sets of screws on the X
carriage.

Preferably the 20mm side... and it should have enough clearance to mount
the standard e3d fan ducts.

Next step is to use the other screw(s) to make a print cooling fan mount
:8ball: