Minor updates, sides of the duct are closed, fan mount is angled for clearance.
I still need to design a clip-in mount for the fan, and a way to mount the duct to the hotends and/or carriage.

Idea: Use a thick-ish band to connect to all three of the hotend mounting holes, thus moving the hotends towards the front of the printer.  We've got a little room there, anyways.  Not sure if the 20mm screws will be long enough, though.