Dual-extruder cooler for the i3.
'tis a thing of beauty.  The blower is canted at a rakish (-5 degree) angle and just slots right in - airflow is uniform and spot on... and everything has like, 1mm of clearance.  Plus it kind of looks like a claw, and makes it nearly impossible to burn yourself on the hotends :-)

You might need longer m3 screws for mounting - the stock is 20mm, dome-head socket cap screws.  On my carriage update, I did sink the nuts in a bit more and it fits for me.  No reason to use the new carriage, although the upper screw mounts won't hurt you.  Odds are good I'll mount something there, just because it annoys me that they're unused.

Paul