Starting work on bg9 design.
Major change in the frame: turn the power supply upside down, use MDF to hide the wires.

Most of the changes are to make the power access panel accessible, and to ensure that noone can touch the voltage.

This commit is in progress, I've updated the z motor mount and U frame.

Still need to modify the base plate, turn the Z motor support into the power supply cover, and extend the back plate to hold the power switch.

Paul