Updated x-end motor
Now both the idler and motor can have a spring, to make the ends into anti-backlash nuts.  Normal usage is one nut; this uses gravity to hold the x-end on.  Advanced usage is two nuts & a spring - which holds the carriage down.  Not recommended for the beginner - a head crash with anti-backlash nuts will break something badly.

Paul